removed circle collision old files
